Restructure the i386 VM code to be able to cope with memory manager requirements.
Basically, this means we have a temporary page which is used for mappings until we can safely allocate memory (vm_map will call vm_map_bootstrap in this case). Once we reach the state in which memory is available, vm_init() should be called which ensures the temporary mapping will be upgraded to a true mapping)

Note that the current code in vm_init() expects that NULL is never an available pointer, and that a single page directory entry is enough to map any memory necessary to finish the VM bootstrap.