amd64: Introduce a 32-to-64 bit relocation trampoline, which is capable of loading and running the amd64 kernel!
The reason we introduce this code is because we want the kernel to be a plain 32 bit ELF file, since anyone can load these. Futhermore, we want to provide the kernel with information we can easily obtain before fully launching amd64 mode (think of memory maps etc), so this trampoline will be the best place to do so.

This will also ease our job writing a bootloader, as it will only have to deal with 32 bit ELF files.