amd64: Fix userland stack corruption caused by invoking system calls
The problem was that the timer IRQ code wouldn't take into account that the CPU-stored %ss:%rsp would be the *userland* version, and it would overwrite the kernel %ss0:%rsp0 later on, causing the kernel to use the userland stack. Once the userland code resumes execution, the subsequent leave/ret sequence would cause the thread to crash.