Clean up block I/O device infrastructure by retiring the 'biodev' hack.
From now on, a block I/O read/write should enqueue an event for the action in the corresponding controller driver. This makes it possible to sort/merge the request queue later on, as well as support command queuing. Some day...

This makes the biodev hack no longer needed, which is a welcome sight. Note that the current implementation expects any block I/O operation to block the caller - this must be made more explicit (it's a side-effect in the current implementation)