amd64: Disable red zones while compiling the kernel.
A 'red zone' means 128 bytes beyond %rsp can be used for temporary data; however, if an interrupt fires during use of such a function, this data would be corrupted. Normally, that cannot happen since an interrupt always introduces a privilege switch, which results in a new stack being switched to.

However, this would give problems once interrupts are enabled in kernel, as functions which use the red zone would get seemingly randomly corrupted...