Move away from an inode-number-based system and move towards file system object pointers (fsop's)
In POSIX, it is normal for every inode to have a number identifying it. While this works great for UNIX-based filesystems, it involves creating mappings and such for other filesystems, notably FAT.

As the notion of mapping a simple number to a file isn't really necessary (since you don't want to enumerate through them anyway), this notion is replaced by file system object pointers (fsop's) - every filesystem informs the kernel of the size, and the kernel will just pass them along.