OpenFirmware: use a platform-specific entry point and clean up ofw_init().
ofw_init() tried to do too much at the wrong time, so axe most of it and leave it to the MD code for the platform. Also, the context in which OFW may be called often needs setting up, so call a known function which should do this and chain through to the entry point.