ata: Fix serious data-corrupting bugs and add more asserts to help catch these in the future.
The problem is that the ATA code will not wait until a request was handled if it is queued. As the request is allocated on-stack in the ata*_read() code, this structure would be read and possible updated causing all kinds of mayhem.

In order to prevent this, copy the request for now. This should be rewritten to use pre-allocated queue entries later.