tracing: Add per-file unique ID's to aid in tracing
This consists of a lot of macro/linker trickery, but the result is that every file will have an unique ID, available using the TRACE_FILE_ID macro (files need to use TRACE_SETUP to be able to use it).

It will be used by the tracing mechanism (which is why it lives here) but also when reporting error codes.