ata: Fix device detection on real hardware
It seems that using the alternate status register (which can be read as often as required as it does not reset itself) is inadequate at this point; it does not seem to contain the correct values. This change alters the code as to only read the status when necessary and storing it for futher use.