fatfs: Fix a bug causing files in the directory that were there, not to be usable.
This happened because we need to keep in mind that we'll only be able to figure out whether a filename fits in the readdir() buffer after we've assembled it. Consquently, if it does not fit, we must reset the directory's inode pointer to the beginning of the entry so that the next read will re-assembly the full file name.