vfs: Correctly handle cache overruns by keeping the caches in sync
The inode and directory entry cache need to know whether an inode is being removed; this change makes them inform eachother so that they will not keep references to inodes long gone. Another problem was that destroying an inode would not unlock it, resulting in a deadlock if the inode was ever to be reused.

While here, use 'vfs_ref_inode' rather than messing with the refcount ourselves.