vfstest: Expand the tests for dentry/inode cache overload.
These will now take into account that the dcache is another ref. While here, update to the renamed VFS fields.

Note that these tests will fail; for some reason, md_reschedule() is called when both dentry/inode cache overflow. This also occurs in real-life use, so I'm committing this anyway so that the issue can be analyzed and fixed seperately.