vfstest: Update the test case
This mainly covers the new inode cache order, but it also adds a check that the entries of the dentry cache should include the files just requested.

Note that this currently fails, which seems to indicate that the dentry cache seems to be purged for no good reason. This needs investigation.