crt/i386: Ensure the caller stack is large enough for the syscall interrupt to fetch all of its arguments
This prevents a page fault when the typical 'Hello world' program is run if it doesn't call 'exit(x);' but rather does a 'return x;' in main(); the syscall code attempts to fetch 6 dwords because that is the most supported. However, there aren't as many dwords on the stack so a page fault is raised instead.

I'm unsure how to fix this in a more generic way without impacting performance too much, need to think about it.