waitqueue: Restructure waitqueues to use an 'add', 'wait' and 'remove' paradigm.
The reason is that waiting on a waitqueue must not construct a waiter; this is in fact quite undesirable because the event may have happened before the waitqueue_wait() function is called. This means the waiter is deadlocked because the event will already have happened and thus waiting should never have commenced.

However, splitting results in the issue that waiting must do nothing if the event is signalled before waiting is used - but as interrupts and syscalls are preemptible, the only way to avoid the race conditions is to use non-premptible spinlocks.