i386: Fix a race when returning to the initially newly cloned thread
The problem is that interrupts weren't disabled, but as esp is already set, any incoming interrupt could overwrite the stackframe constructed by our syscall handler. As we strongly depend on this structure, we could restore invalid registers, causing chaos 'n mayhem all along in the destination thread - where, of course, the problem would manifest itself much much later.

Note that this magic isn't strictly necessary because we could craft a special stack for the restore handler - but this is awkward to implement now and hasn't really got any benefits for now.