i386: Clean up the syscall/clone return code to make it more readable
There's no reason why we should create a stackframe, as simply pushl-ing the values on the stack one by one is easier, faster and easier to follow. Also, we can re-use a large part of the syscall return code after doing a clone, making it clearer what is happening there.