i386: Fix an incredibly annoying bug that sometimes causes %ds = 0 on syscall return
The problem was that %ds wouldn't be restored if the context restore code found that it would return to kernel code - this is wrong, because there is no real guarantee that %ds was any specific value. The result is that the 'iret' would notice that %ds is ring0-only and it would be set to null; the userland code would crash as it can't access its own data anymore.

Note that this can only happen with preemptible syscalls, as the problem only occurs if there is a scheduler interrupt between the syscall irq and the 'push %ds' - this would cause the wrong %ds to be stored and restored later.