kernel: Add a rescheduling-flag per thread and set it when something more important was woken
This is designed to keep the system from spending all time in the idle thread when an IRQ resulted in a more important (currently any other) thread could run.

The I/O speed introduced by this change is very noticable; profiling has shown that most time was spent in the idle thread, which would generally just wait for the timeslice to expire while much more useful work could be done instead.