kernel: Fix a hang, crash, double fault or just plain reboot when more than a few IRQ's occur relatively quickly
The problem is that IRQ's are typically handled in the idle thread: threads go to sleep when they wait for an IRQ and the scheduler wakes them. However, once the IRQ is serviced, it'll use the idle thread's kernel stack - but, since the idle thread is never run after this (as the original thread can continue), it won't be able to clean up its stack. And it gets worse quickly, because if the IRQ's occur quickly enough (which is rather trivial, for example during 512-byte disk I/O operations), the stack just keeps growing larger, larger ... until it finally runs out and overwrites something critical or results in a double fault.