kernel: Fix a serious bug in the waitqueue_wait() causing panics in the scheduler code
The problem is that there is a race in the following code:

	thread_suspend(t);
	spinlock_unlock_unpremptible(&w->w_lock, state);
	/* Race here */
 	reschedule();

The result is that the thread would be marked as 'suspended' while it is active, but the scheduler IRQ could kick in during the race described above. The main problem is that 'thread_suspend(PCPU_GET(curthread));' is racey and is only actively used here - this change avoids the problems that it causes but this is not a definite fix.