i386: Fix a bug that may result in a cloned thread returning to the wrong address
It turns out that a system call will always use a clean kernel stack upon entry; this is the case because it is not possible to perform a system call from any other context. This implies that we only have to copy the first few bytes that were used to enter the kernel (syscall_int sets these up)

It is plainly wrong to rely on the parent's %esp as it will not point to the initial stack anymore if there was a pre-emption in between: the %esp will point to an inner-stackframe which is no longer valid by the time we return to our syscall code.

The result is that we could execute the pre-empted code twice; this would go unnoticed because we would just adjust the thread's stack again, but in the worst case it sometimes led to hard-to-understand problems since we shouldn't have gotten there with the current thread [1]

This change restructures the code so that we only restore the bare minimum kernel stack needed to return to the thread's caller; this is much more efficient and has the main benefit of not messing things up if the clone is pre-empted.

[1] This reasoning is deliberately vague; I'd need a crystal ball to figure out the exact course of action :-)