kernel: Introduce pipe handles
Note that the behaviour is different from POSIX in regard that our pipes will never block - if blocking is desired, the operation should be attempted and followed by a sys_wait(..., HANDLE_EVENT_{READ,WRITE}, ...). This makes our pipes less efficient than the POSIX ones, which may be fixed with a 'flags' argument to read() some time?