i386: Fix SMP support
This introduces a specific interrupt source for the APIC. It also moves the SMP code to the front because we cannot re-attach interrupts, nor should we want to: we should determine whether to use the PIC or APIC before actually initializing devices.

AP's will now panic because the idle thread isn't initialized correctly, this is for later.