kernel: Fix a big whoops causing threads to be scheduler on multiple CPU's at the same time
The problem is that threads that can be run always remain on the runqueue, but the scheduler doesn't check if a thread is active and assumes it is not when it encounters a thread on the runqueue. This would cause multiple CPU's scheduling the same thread, which blows up very fast.