kernel: Rework the scheduler to avoid a nasty context corruption bug
The problem was that the context switching mechanism would set the old thread as inactive _before_ the switch to the next thread was completed; this could cause it to be picked up by another CPU before it was actually done with the switch to the new thread.

This often manifested as two CPU's deciding to run the same thread, in which they would corrupt the new thread state.

The fix is, obviously, to set a thread as inactive once the new thread itself is fully activated. This requires a change to md_thread_switch(), since we need to obtain the thread where we came from - this is not necessarily the previous thread when we switched (as returning from 'md_thread_switch(new, old)' does not necessarily mean 'old' re-actived us, it may be a different thread)