kernel: Rework thread termination to prevent races on SMP systems
The main problem is a race between handle_signal() and handle_wait() - the latter could wake up immediately after being signalled. The thread subsystem depends on the thread exit process being completed _before_ any waiters get a chance to avoid 'thread is destroying itself problems (a thread can't destroy itself because it needs stack).

This removes the 'terminating' thread flag, which does not contribute anything (no special knowledge is needed when a thread is terminating) and was quite error-prone because it is okay for a thread to sleep during termination because it may be suspended on mutexes and such.

Sadly, this adds thread-specific knowledge to handle_wait() - this should be reworked in a specific handle operation, but that is for later.