kernel/ahci: Do not issue an interrupt after completing a PRDT entry
The data may still be in transit, and there's no guarantee that it is available
to the CPU (when reading) - and note that this only covers a piece of the data
(as it's a single PRDT entry) so we don't even care about it once we start
merging requests...

Furthermore, we'll get an IRQ once the 'command issued'-bit goes from 1 -> 0,
which signals that the _command_ is completed and thus all its associated PRDT
entries, which is what we actually want.

This was found when testing the AHCI driver in QEMU, which doesn't implement
this bit at all (both FreeBSD and Linux do not use it) - removing the bit makes
the driver work on these emulators and doesn't make a difference on actual
hardware.