kernel/amd64: Rework syscall enter/exit code to allow restoring full context
We'll now use the same stackframe layout for both syscalls as well as
interrupts/exceptions. This allows us to use the same return path for all
instances, which is convenient.

This is necessary for syscalls like execve(), which need to restore the
entire context in the new executable's situation.

While here, some minor changes (i.e. pid numbers during userland exceptions,
that sort of thing)

Inspired by: FreeBSD