kernel: Rework up device attachment
This differentiates between attaching a device (which attaches a most a single
device) and a bus (which can attach multiple new devices on that given bus)

The previous code attempted to do this in a single, unreadable function,
resulting in hacks in anything that wanted to probe for and attach child
devices (PCI, USB). This makes things easier to follow and gets rid of the
hacks.