kernel/usb: Fix OHCI bulk transfers
With the C -> C++ transformation, not all field members were properly
initialized. The result is that the 'toggle data0/1'-bit is generally always
true, which means we will not send the correct DATA0/1-packet.

This causes the packet to be blatently ignored by the device, which is
extremely unhelpful. Note that QEMU's USB support does not adequately
emulate the DATA0/1-behaviour, so it worked there.