kernel: Store the boot arguments during early MD startup
The main problem is that kmem_unmap() may destroy mappings we really need, such
as pagetable mappings or the initial KMEM_MAPPING*.

This is hard to avoid, so it is much easier just to avoid the problem. Plus, by
the time we get to cmdline_init(), it would be pure luck whether or not something
claimed the memory where the cmdline was stored anyway.