kernel: have exec_load() take own dentry_ref() instead of transferring from caller
The new dentry cache (about to be committed) correctly discards dentries if needed,
which showed that init-userland.cpp didn't expect the dentry ownership to be
transferred.

It makes sense to have exec_load() just take the ref as needed (the entire
concept of transfering refs is confusing and error-prone) so rework the code to
reflect this.