mb-stub: Make this build with clang
Major issue is that GCC 6.x shipping with Debian default no longer builds
something that works (likely because it generate PIC code?). We need to
ensure no SSE is used or this'll crash (perhaps we should just enable
it instead?)

Also, clang names .rodata differently, so cope with that in the linker
script.