kernel/vm: Add Copy-On-Write functionality to the VM subsystem
This turned out to be a huge pain, mainly because this is extremely hard to
debug - I'm thinking that deadlock-debugging should be added to help with such
things.

This change adds sensible locking to VM_PAGE (vmarea_t still needs this!) - for
this to work, VM_PAGE needs to be aware of the vmarea its in (this should
replace the PRIVATE flag).

The code isn't as clean as I hoped it would be, but improvements will likely
happen at some point.