kernel/amd64: Prevent corruption of %r13-%r15 after first system call
It is astonishing to see that this was broken for >1 year without too many
problems - once I start building RTLD with -O3, seemingly random crashes would
occur because %r15 was clobbered.

Ironically, the fact that we fully restore the thread state after the first
system call is wrong - there is no need to do that. I'll need to dive in further
to see how I can avoid this (we only need the flag after returning from a
sys_execve() syscall)