libm: Introduce libm, the math library, obtained from libmusl
We need a libm as a prerequisite for libcxx, which is in turn needed to build LLVM/Clang and all other things of goodies.

Special thanks to Ed Schouten <ed@nuxi.nl> for suggesting this particular library, which seeing how quickly it could be integrated likely was the best choice.

Note: this has not been tested in any way.