kernel/x86: Implement interrupt levels and polarity
We used to assume that everything was edge-triggered, high-polarity - this is
wrong because PCI interrupts are typically level-sensitive, and the ACPI MADT
details which mode is to be used, which we ignored.

The result is that when a level-sensitive interrupt is triggered while we are
already servicing one, we will miss it since we mask out all interrupts and
after unmasking, the interrupt line never changes.

This bug caused the AHCI driver to miss an interrupt if multiple requests were
in progress and only one of them was ready once the interrupt was being
processed - the hardware would trigger the next interrupt but as we wrongly
expected an edge-triggered interrupt, we would not see it, resulting in a
soft freeze.