rtld: fix importing global symbols from a common library
Assume you have executable 't', library 'libfoo.so', and 't' depends on
libfoo.so, and both depend on libc.so - both 't' and 'libfoo.so' try to
use, say, 'stdout' from libc.so.

In that case, 't' would import 'stdout' from 'libfoo.so' as it did not correctly
check whether the symbol was *defined* in the first library it saw - it would
assume it did. The end result was very weird crashes and a lot of head-scratching...