kernel: Fix seemingly random crashes when spawning new processes
This whole mechanism is just a mess - we need to allocate a separate vmspace
for it since we need the NO_CLONE flag, and then we specifically allocate pages
for it.

The problem was that vmpage_create_private() adds the page to the vmspace,
which will thus be freed by vmspace_destroy(). The result is that we end up
freeing the page twice.

This would normally not be a problem, but alas there is a race between
destroying the process and exec()'ing something in the new one. This means
we could allocate the page that was destroyed by vmspace_destroy(), only
to find it unmapped by kmem_unmap() when the rest of the process dies.

This was broken since the VM rework; the intention is to get rid of the
process info in the near future and replace it by the auxiliary vector stuff
that the x86 ELF ABI suggests - this would make things a lot easier.