kernel/vm: Do not attempt partial cleanups of vmspace
vmspace_destroy() was never invoked during ordinary process cleanup, which
results in a resource leak. This change hides vmspace_cleanup() so that we
can only clean up vmspace's fully.

This means that we will only destroy the vmspace completely on process exit
instead of upon termination. The net effect is that zombies are more expensive
resource-wise. For now, let's accept this.

While here, ensure 'x = new ...;' is in sync with 'delete x;' (do not call
kfree() directly)