kernel/vm: Fix unpredicable behaviour due to refcount bug and restructure this code
The problem was that, upon promoting a VM page from COW to writable, it
wouldn't be removed from the VM area it belongs to. This causes problems later
on:

1) The refcount would be incorrect as the VM area still has a reference to it
2) The page would be copied during a VM space clone as the VM area still finds it

Sadly, there were no checks on the refcount - and even worse, as the VM page would
be removed, we'd end up with a dangling pointer causing heavoc later on (links to
linked paged were most common, but things would occassionaly be completely corrupted
as the pointer did not actually contain a VM page)

The solution is to deregister the page before Deref()-ing it; while here, a lot of
restructuring was done and assertions were added all over the place to aid
debugging. For now, I'll leave them in as this code is really horrible to debug as
every line is 'hot' and I was almost always looking at damage already done...

This unfortunately also contains two libc rename actions.