kernel: Prevent deadlocks during wait() when multiple processes invoke wait()
Again, it was a sad decision to use a single semaphore; this avoids the deadlock
by using a semaphore per parent, but that still is not enough as multiple threads
within the same parent could wait.

I should invest time in getting a sensible condition variable-like
synchronisation primitive in place.