build: Remove dependency on x86_64-ananas-elf-... symlinks
This allows us to just have 'clang' which will specify the target as needed.
This makes things a lot cleaner and will pave the way for distcc later on.

Note that this consisted of a lot of specific overrides to get CMake
to actually use the _correct_ tools. I am still not convinced this is
fully correct, and this will likely need work in the future.