kernel: Do not thrash the interrupted calls return code if a signal handler is to be invoked
This was observable by starting cat, hitting ^C and pressing enter - we can't yet interrupt
read(2), so after the enter the signal magic would happen and we'd kill cat(1) and the
signal handler of dash would run, onsig().

It concludes that nothing worthwhile has happened so it in turn returns, and it expectes %rax
to contain the exit code of the waitpid(2) syscall - which was actually the
result of sys_sigreturn(), which always returns zero.

dash gets confused and calls waitpid(2) again, as the actual pid of whatever
returned was lost. This change fixes that.