build: Invoke 'clang++' for C++ building/linking instead of 'clang'
This removes the need to explicitly add '-lunwind -lc++' to the final C++
linking command, as 'clang' is the C driver and assumes we don't want to
link C++ libraries.

While here, pave the way to building using distcc by introducing the ..._D
variables that could have 'distcc' prefixed to them. This needs further
testing, though.