rtld: Fix infinite loop when DT_RELAENT is not present
This value is not mandatory for shared objects, and thus may not be present at
all - we'd use the default of zero in such a case, which meant we'd be
re-processing the current entry over and over.

This changes removes support for different DT_...ENT sizes - the code should
have checked that at least the fields we require are present, and it didn't
even do that. Let's just axe support for this 'feature'.