kernel/amd64: Clean up SMP code
We'd have acpi-smp.cpp as the settings necessary to setup SMP would be parsed
from ACPI or the MP table. Support for the latter was axed a long time ago as
I've never seen a table with correct contents - and since we depend on ACPI
anyway, let's forgo the old structure.