kernel/hda: Rework HDA routing code
This also contains several other improvements, such as:

- Add support for audio selectors
- Add 'support' for volume knob widgets
- Set power state to D0 upon attach
- Failure to establish a routing channel ignores it
- Minor code cleanups

With these changes, audio can be played on my Dell Optiplex 160, which contains
several audio selectors in the audio graph.

Unfortunately, VirtualBox seems to receive all audio correctly but does not
play anything. This still needs more work - a hint I got from a developer is
to make sure SD0 is used for input and SD4 for output - but I think this is
already the case. Either way, this needs more work.