kernel/x86: Rework IO-APIC interrupt routing
We'll map the first 32 IRQ vectors now - this needs work: we should only map
the ISA IRQ's and do everything else dynamically.

This fixes masking/unmasking of IRQ's using the IO-APIC: we would implicitly
assume a 1:1 relation between IO-APIC pin# and IRQ#. This meant we'd mask
pin 0, ExtInt, which is intended as a 'virtual wire' to route interrupts to
a PIC-like schema. This would mask the PIT and everything else, and ensure
we could only receive ISA interrupts.

A big thank you to John Baldwin of FreeBSD fame, who was kind enough to explain
the concept of the 'virtual wire' to me!