Re-license to the Zlib license
The main reason is that beer-ware is sketchy in some ways: it is not studied as
well as the more common licenses, and this may hamper adoption.

As one of the main goals of Ananas is to encourage code re-use, it made more
sense to use a more commonly understood library. The main contestants were the
MIT license and the Zlib license; I've opted for the latter as it is more
liberal.

I would like to thank Arnoud Engelfriet for his helpful suggestions when
discussing license requirements and expectations.

While here, remove the 'vim' lines as these are no longer necessary.