kernel/vm: Force truncated dentry-backed mappings to a full page length
This fixes a SIGSEGV when:

- An application maps something RW from the last part of an inode
  Let's say the pointer is 0x1000, which spans offset 0..0x900
- The inode's length is not a multiple of PAGE_SIZE
  As the inode is only 0x900 bytes long
- An address beyond the inode's offset is written
  This could be address 0x1900, which is just outside

This would cause a SIGSEGV because the mapping would be limited to the inode's
length and not to what actually is the memory backed. This is wrong: mapping
lengths should always be a multiple of PAGE_SIZE as these are the addresses
that we can fault.

Found when strip(1)-ing /bin/sh, which exhibits this behaviour.