Clean up LLVM docker image construction
Instead of downloading the entire LLVM tree, we now copy it from our current
repository.

The current repository also has the patches integrated (that was the point
of having a specific repository) so we can get rid of all the patching in
between.

While here, remove all 'customisation' which would mess things up anyway
and just hardcode the values we need.

The idea is that this makes changing things much easier and maintainable.