kernel: Rework the block I/O layer
The design is greatly influenced by NetBSD, which in turn is based on the
original UNIX design. This mainly brings us proper locking, which will
be important once we start delaying writes.

Some corners were cut, mainly:

- We really need condition variables to make the locking work
- BIO's are not assigned to inodes, which means we tend to
  saturate the buffer mutex
- Delayed writes are not yet supported
- No effort was made for combining I/O operations
- Data allocation is still a mess

... so this is slower than what it used to be :-(

On the flip side, this should implement proper locking, should be correct
and be a good base for future improvements.