kernel/bio: Use condition variables
This seems to improve performance substantially - it also addresses
that when we are waiting for a block in the cache to become usable,
we should wait until the _owner_ releases it, not when the _data_
is available. This could become a nasty race condition, so I'm glad
it's eliminated now.

Again, the locking is almost a copy of what NetBSD does (if it works
for them, it'll work for us)