kernel: fix race between thread destroy and the scheduler
Given enough CPUs, one core could be calling Process::WaitAndLock(), which will
destroy the thread belonging to the process. However, said thread could still
be active on another core in md::thread::SwitchTo() - which means we'd blow up
as we would remove the kernel stack of the thread.

This just waits until the thread is properly released by the scheduler before
we free the final resources.