kernel: Always save/restore x86 FPU state
Since SSE is mandatory on x86_64 and it is used all over the place in userland,
let's just always save/restore the context.

For now, don't care if it is a kernel thread or not, which may bring us closer
to allowing FPU use within the kernel, which is in turn needed for most C++
runtime libraries. This will likely make interrupts a bit more expensive,
though once it is enabled there, but that is for later to worry about.

Note that Linux switched to non-lazy FPU save/restore (called eager) since
early 2016.