libc: Integrate libpthread into libc
It's arguable that having a seperate libpthread is a thing of the past as libc
predates libpthread. It makes things a lot easier just to merge them (i.e. less
messing around with dependencies etc)

It is worth pointing out that Musl also merges these, so there shouldn't be
much harm doing the same.