kernel: Introduce fixed-item length memory pools
Pages will be allocated as necessary for these; the main benefit is that
these are far more efficient than the bitmap kludge the BIO code used
and we can use them through the kernel, yielding common functionality.

Note that the pools will never free unused pages; this will be added
once needed.