vm: Remove linked VM pages by keeping a vector<> of pages
The main reason for linked pages is that an INode owns the pages, yet a VMArea
needs to know which pages are in there for things like COW.

Instead of making things overly complicated, just using a vector<> to find the
pages makes things a lot easier.

Note that this disables COW for now until I get around to reimplementing it.