kernel/vm: Allow VMPage's to be shared without a linked page in between
This allows to use the refcount to share pages instead of an in-between 'linked
page'. This simplifies things greatly and since sizeof(VMPage) == 150 the last
time I checked, it actually saves memory since util::vector has way less
overhead than that.

This may be the best example of "a potentially good idea gone terribly wrong"
:-/