kernel/vm: Clean up the VM code
This removes unused flags and data members that aren't needed. It also removes
the lookup through VMArea's pages for a DEntry page as that simply cannot
exist when we are faulting (we need to fill out the page!)

There are parts of COW in here which are disabled; finishing that up later.