rtld: Fix skipping extra argv[]
As the argv[] and envp[] aren't passed to the executable from here, the result
is that we may miss auxv[] arguments - if no environment is passed, the rtldj
would crash as we can't find the interpreter base.

This change removes the progname skip which is incorrect (it is counted in
argc) and adds asserts to ensure we have our proper information available.