kernel: Properly support edge-triggered interrupts
ISA are edge-triggered, and when they occur in rapid succession it is very
likely that we mask the source, causing us to miss the interrupt. Thus, we
must never mask interrupts, but rather keep note that they triggered when
we were handling them. This allows us to acknowledge the interrupt straight
away without having to worry about stack usage exploding.

The strategy used here is inspired by Linux, which lets the interrupt arrive
and sets a flag to indicate that the current handler has to be run again. This
flag can only be set whenever we are in the OnIRQ() handler because interrupts
are disabled.