awm: Improve performance by adding an update rect
This reworks the 'need update' boolean to an 'update rectangle' which is what
will get updated/drawn this frame. The mouse cursor has been updated to
properly fit into this - this prevents having to copy lots of bytes for no
benefit.

Note that further optimizing memcpy() would be very beneficial here.  Also, the
mouse cursor is hastily drawn and may not be very aesthetically pleasing.