kernel: Remove usage of ananas/types.h in the kernel
These types are shared between userland and kernel, which means that
ananas/types.h will not publish the type (it will be prefixed with __).
This change adds kernel/types.h which includes ananas/types.h and
publishes the required types without the prefix.

This is still messy and needs further cleanups.