kernel: Rework the child/parent process infrastructure
If a process has no parent (which occurs during startup), it will be changed to
init during the startup process. This ensures init can clean up any zombie
process, including kernel processes, which makes things a lot cleaner.

This also implements waitpid(pid, ...) for pid > 0, where we want to wait for a
specific process to terminate. This is necessary because /etc/rc mounts /ankh,
which is needed for the getty's to run. Because there are multiple kernel
processes that terminate once init is run, we'd continue immediately without
properly waiting for our init script.