kernel/vfs: Do not cause an unnecessary deref when a symlink cannot be resolved
This would result in panics in dentry_deref_locked as the refcount would be
zero once we got there, where in fact it would have been dereferenced
unnecessarily.

A lot of the code has been refactored by this commit in an attempt to make it
easier to understand and maintain.

Clean up resolve/follow logic