kernel: Clean up thread/process/scheduler interaction
The main benefit is that the scheduler will make decisions based on
oldThread->t_state - i.e. if we want to suspend the current thread, it is
sufficient to set t_state = Suspended and call Schedule(). This is based on the
approach taken by Linux, which really helps makes things easier.

A disadvantage of this approach is that we can't (reliably) sleep on thread
termination, as setting t_state = Zombie (which is needed for WaitAndLock) will
cause the scheduler to never pick the thread up again. There may be ways around
this, but I prefer to keep things simplier and sub-optimal as this is a very
complex code to begin with and I'd prefer to get it right (which remains to be
seen)