kernel: Fix a race condition in Mutex::Unlock()
Given:

    void Mutex::Unlock()
    {
        KASSERT(mtx_owner.load() == &thread::GetCurrent(), "unlocking mutex %p which isn't owned", this);
        mtx_owner = mutex::unlocked;
        // <--- mutex can be destroyed here as it is no longer locked
        mtx_sleepq.WakeupOne();
    }

In the VM code, a lot of locking, creating, destroying and re-using is going on. This means that
you can get at the comment line stated above, and the VMPage would be delete'd. However, the
original thread unlocking the Mutex was not done with it. And since the memory is so hot, it's
likely it does all kinds of nasty things.

The solution implemented here adds an extra flag to ensure the mutex cannot be
re-acquired while we are signalling others (which should never happen when we
are about to delete the memory in the first place), and while we are waking up things, we
disable the interrupt flag to ensure we cannot be pre-empted away -
SleepQueue::DequeueWaiter() needed to disable them regardless so this is not a big loss and
it makes things clearer.

This took a lot of time to pin down - without the fix, the system would hang due to a corrupted
spinlock (until the lock debugging commit improved things) or panic due to garbage data otherwise...